// src/app/models/customer-address.ts
export class CustomerAddress {
    id = -1;
    addressable_type: string;
    addressable_id: number;
    original_address_id: number;
    name: null;
    first_name: string;
    last_name: string;
    company: string;
    address1: string;
    address2: string;
    city: string;
    country: string;
    province: string;
    district: string;
    zip: number;
    phone: number
}